pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "people-scout"
version = "0.1.0"
description = "AI agent that generates person intelligence reports from a Twitter handle"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Scout Team" }]
keywords = ["ai", "twitter", "intelligence", "selanet", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.22.0",
    "ruff>=0.8.0",
]

[project.scripts]
people-scout = "people_scout.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/people_scout"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
